refactor(utils): move backend utils
