+++ src/backend/mod.rs
@@ -1,2 +1,5 @@
 pub mod canvas;
 pub mod dom;
+
+/// Backend utilities.
+pub(crate) mod utils;
